[fix] database acess

+++ src/main/java/com/beerhouse/repository/BeerRepository.java
@@ -3,5 +3,5 @@ package com.beerhouse.repository;
 import com.beerhouse.models.Beer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BeerRepository extends JpaRepository<Beer, Long> {
+public interface BeerRepository extends JpaRepository<Beer, Integer> {
 }
